Use pinv to calculate LLR.
Ok, i'm looking at your code, and it seems that this is happening
because several of the eigenvectors produced by dmaps have many
near-zero entries. So, the median median distance calculated in the
first 30 lines of local_linear_regression is also near zero, and you get
points "linearly" reconstructed from only a single neighbor. This would
normally be fine (the reconstruction would be just that neighbor's
value, with an undefined slope, and we don't actually need the
slope(beta)  to calculate FX), but sometimes there are a couple of
neighbors, and this somehow messes up alpha too.

This problem gets solved if you use the pseudo-inverse to solve the LLR
rather than backslash (
A = pinv(Xx2*Xx) * Xx2 * Y;
rather than
A = (Xx2*Xx)\Xx2 * Y;
)
But, still, having lots of X points essentially on top of each other in
the LLR code leads to very small eps (in LLR, not dmaps), and so you get
a LLR where each point is essentially unaffected by neighbors, and so
the residuals are artificially high.

So, you can try to bias the calculation of eps inside the LLR code to
ignore zero-ish distances when calculating a median (or mean? another
option). But then, you get artificially low residuals, since eps
suddenly jumps up, and you get undue influence of that giant cluster of
zero-X predictors, so the FX gets dragged towards zero. (This happens
especially if you use the mean of K values to find eps).
towardZero.png

So, using pinv instead of \ eliminates the NAN residuals, but the
underlying problem of too-low eps internal to the LLR code is still
there, meaning that you'll get lots of artificially low residual values.

I'll continue thinking about this--Ideally, you'd change your dmaps eps,
so that your dmaps vectors don't have large constant segments, but I
think the LLR code should be able to handle such cases gracefully.